Fixed #20
I also did some general cleanup of the use of variable names. I left some stuff, like the use of `names` instead of changing it to `qNames` as I'd done elsewhere because there was a lot of stuff with regex being done and it felt like it no longer exactly reflected the query names or sequence names, it was just grabbing a shorthand or something.